+++ addOns/exim/src/main/java/org/zaproxy/addon/exim/har/MenuImportHar.java
@@ -19,17 +19,25 @@
  */
 package org.zaproxy.addon.exim.har;
 
+import edu.umass.cs.benchlab.har.tools.HarFileReader;
 import java.io.File;
+import java.io.IOException;
 import javax.swing.JFileChooser;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.model.Model;
 import org.parosproxy.paros.view.View;
+import org.zaproxy.addon.commonlib.ui.ProgressPane;
+import org.zaproxy.addon.commonlib.ui.ProgressPaneListener;
 import org.zaproxy.addon.commonlib.ui.ReadableFileChooser;
+import org.zaproxy.addon.exim.ExtensionExim;
 import org.zaproxy.zap.view.ZapMenuItem;
 
 public class MenuImportHar extends ZapMenuItem {
 
     private static final long serialVersionUID = -9207224834749823025L;
+    private static final Logger LOG = LogManager.getLogger(MenuImportHar.class);
     private static final String THREAD_PREFIX = "ZAP-Import-Har-";
 
     private int threadId = 1;
@@ -53,8 +61,30 @@ public class MenuImportHar extends ZapMenuItem {
                                     public void run() {
                                         this.setName(THREAD_PREFIX + threadId++);
                                         File file = chooser.getSelectedFile();
-                                        boolean success = HarImporter.importHarFile(file);
-                                        if (!success) {
+                                        ProgressPane currentImportPane =
+                                                new ProgressPane(file.getAbsolutePath());
+                                        int tasks = 0;
+                                        try {
+                                            tasks =
+                                                    new HarFileReader()
+                                                            .readHarFile(file)
+                                                            .getEntries()
+                                                            .getEntries()
+                                                            .size();
+                                        } catch (IOException e) {
+                                            LOG.warn(
+                                                    "Couldn't count entries in: {}",
+                                                    file.getAbsoluteFile());
+                                        }
+                                        currentImportPane.setTotalTasks(tasks);
+                                        ExtensionExim.getProgressPanel()
+                                                .addProgressPane(currentImportPane);
+                                        HarImporter harImporter =
+                                                new HarImporter(
+                                                        file,
+                                                        new ProgressPaneListener(
+                                                                currentImportPane));
+                                        if (!harImporter.isSuccess()) {
                                             View.getSingleton()
                                                     .showWarningDialog(
                                                             Constant.messages.getString(