+++ addOns/exim/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Reduce logging and display a warning dialog when unable to read files being imported (Issue 7081).
 - Promoted to Beta.
 
+### Added
+- Importing a file of URLs or HAR is now displayed in the progress panel provided via commonlib.
+
 ## [0.0.1] - 2021-12-22
 
 - First release.